Refine code.
